requirements.txt
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
packaging>=23.0
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
colorama>=0.4.6

